use FormControl<string> for numerics, disable submit button
FormControl<string> allows placeholder text to be shown. There's no need to use FormControl<number>.
Field values are still saved as numbers if the input type is number in
the template.

Even if the input type is text and the field value is a string, it is
still converted to number by mongooose or mongodb when the document is
inserted (as long as the value inside of the string can be converted to
a number)